Made the body of home page into a flatlist
Was able to add a banner and an example listing/product, all into a flatlist of the body of the homepage. Everything seems to be going well. Instead of replicating the business webpage (mobile/phone view), I'm trying to replicate the layout/design of the eBay app, as it is more simplistic and a better design (and a design a lot of people can relate to if they were to use the eBay app) compared to the web view of our website.