Added comments for styles and a simpler listing
I added more things in terms of layout and design and added helpful comments for the styles. I was able to figure out a way to have the navigation list open without changing the styles and layout of the homepage. Going to work on fixing up heights and widths to be based on device dimensions, creating the listing to be a pressable component, and creating the navigation system to different pages of the app.