Setup MySQL and GET Request with fetch
Was able to set up a python-flask webpage to act as a request API to call the MySQL Database that holds all product data (and will eventually hold customers and orders). This allows the products to not have to be saved and loaded on the app itself; everything is connected through the internet/requests and HTTP calls. Will work on making the FlatList components to use this new system of requests, and figure out how to save the pictures (thinking of using a photos database that saves the photos as links, and the MySQL DB will hold the links to each specific image).