Images now saved through Imgur and MySQL
Was able to start storing images into an Imgur account, and the MySQL database now stores the link to those images, which is all called through the fetch/request process of the app. Switched all components of the app to use this database and links to images instead of manually-downloaded images and information. Will continue to work on the product page information, as well as writing things more efficiently (the process of requesting/creating the array of products can be a simpler process, such as only being done in the full "App" component of the app (the "extern default function App()" section/component)). Other than that, it's making good progress!