Created simpler design for the navigation list
Instead of the navigation list opening from the side of the screen, distorting all of the current components on the page, I learned of absolute position, zIndex, and opacity to make the navigation list appear on top of the current page, instead of pushing itself through the left side of the screen. The user is able to close the navigation list by pressing anywhere on the screen (other than the navigation list), similar to various mainstream apps. This allows for a more efficient system of viewing pages and their components, instead of trying to create a way of distorting the components when the list is opened. I will work on making this process more efficient (still a bit glitchy and resets the home page when the navigation opener is pressed), and will also work on the product page information and layout.