Made Navigation List more efficient and clean
Developed a more efficient method of showing/hiding navigation list - no need for two if statements anymore for rendering the pages. Instead, the components and their styles change based on if the navigation list opener/button is pressed, making the transition more clean and reducing amount of lines of code needed. Will now move on to animation for the navigation list, and adding more information for the product page.