Learned of Text Input and added it in
I created a search bar and a search result page, using similar components of the Category page. The search page allows for the search query to be edited while continuing to narrow down a users search. I created an algorithm for splitting the search query into a list of the words used, and searches for each word within the categories and titles of every listing, looking for a match in wording for displaying search results. Will now move on to adding more information for the products, as the main "navigation" side of the app, in terms of navigating to different product and category pages, is complete (may add some more things in these terms if I come up with any other ideas for it, apart from the "cart" and customer account pages).