Significant change made to categories/search page
I combined the search and category pages together to make everything more efficient and to make the app more minimalistic/simple, as the category and search pages were very similar (just the difference of a title). The title for the category pages instead will go into the search bar of the page. Also, to make the layout of the title component for each page consistent and easier to create, each page will have the same title component- the navigation button or back button, Capelli logo, and account & cart buttons (may change in the future if more buttons are needed). This allows the navigation between pages to be smoother and look much more consistent and attractive than what I used before. I also created/painted a blue account and blue cart png that are used for their respective pages (cartSelected and accountSelected images/variables). Overall, this change is extremely significant and makes the front-end work much simpler/efficient, and I can soon work on the back-end of the app - account and cart management. I tested the app on all emulators and most of my family's and friends' phones, and it's looking great!