Was able to get accounts with fetch request working
Created an account login page, and established to connection with the submit button with a URL I added for account management. Database for users has been created, and the posting of data to the specified URL is working correctly. Currently working on the backend - checking if user/email already exists when a new user signs up, and will return a response/value for the app to portray to the user, or for the app to do something in terms of what the response is. Will also create two separate pages for logging in and signing up.