Updated animations and login/signup pages
Animations between pages, especially using the navigation list, are much more clean and efficient, instead of a simple fade animation. The fade animation is only kept for the product, cart, and account pages to keep the animation of the icons to look more natural and attractive/interesting. The back-end of the site has been updated to provide a response for both signing up and logging in. I'll work on making the app save the current logged-in user, and afterward work on creating a cart of items for them, as well as a way to add certain items to their cart. It's coming along great!