Created Account page and navigation between pages
Created the system of navigation between the signup, login, and account pages. Works very well, but I may encounter some bugs with it as I'm sure there is something missing that may go wrong in the future. For now, will work on putting/storing more info for the users account, such as address, orders, etc, as well as creating buttons/pages for these user account options.