Created quantity system for product and cart pages
System for keeping track of quantity is working well, and the text Input components for the quantity entry look good. Will work on completing the changeQuantity function to automatically update the quantity of the product in the user's cart whenever the user finishes editing the text Input in their cart page. Afterword, will work on adding and deleting items from the cart, and adding a description section for the product pages.