Created address editing system
Address system is complete, in terms of viewing and editing a user's address. I'm currently finishing up the addition of a new address (just need to finish the backend), and afterward will start working on the system for payment systems, and on to other various settings a user can control and change to finish the account/user page.